fix: update readme and remove contracts parts from frontend examples
