+++ src/components/web3/WalletList.tsx
@@ -3,7 +3,7 @@
 import { WalletType } from '@bgd-labs/frontend-web3-utils';
 import React from 'react';
 
-import { useStore } from '../../store/ZustandStoreProvider';
+import { useStore } from '../../providers/ZustandStoreProvider';
 import { WalletListItem } from './WalletListItem';
 
 export const WalletList = () => {
