+++ src/providers/ZustandStoreProvider.tsx
@@ -4,27 +4,7 @@ import { createContext, type ReactNode, useContext, useRef } from 'react';
 import { create, type StoreApi, useStore as useZustandStore } from 'zustand';
 import { devtools } from 'zustand/middleware';
 
-// combine zustand slices to one root store
-import {
-  CounterSlice,
-  createCounterSlice,
-} from '../counter/store/counterSlice';
-import {
-  createTransactionsSlice,
-  TransactionsSlice,
-} from '../transactions/store/transactionsSlice';
-import { createWeb3Slice, IWeb3Slice } from '../web3/store/web3Slice';
-
-export type RootState = IWeb3Slice & TransactionsSlice & CounterSlice;
-
-const createRootSlice = (
-  set: StoreApi<RootState>['setState'],
-  get: StoreApi<RootState>['getState'],
-) => ({
-  ...createWeb3Slice(set, get),
-  ...createTransactionsSlice(set, get),
-  ...createCounterSlice(set, get),
-});
+import { createRootSlice, RootState } from '../store';
 
 // provider with zustand store https://docs.pmnd.rs/zustand/guides/nextjs
 export const ZustandStoreContext = createContext<StoreApi<RootState> | null>(
